refactor: 폴더 구조 변경 및 컴포넌트 파일 축소
- type, index.ts 삭제
- Component 관심사별로 크게 구분하도록 폴더 구조 변경
