--- src/components/Auth/@subComponents/AuthContainer/AuthContainer.type.ts
@@ -1,5 +1,0 @@
-export namespace Props {
-    export interface AuthContainer {
-        children: React.ReactNode;
-    }
-}
