--- src/components/Auth/@subComponents/AuthContainer/AuthHeader/index.ts
@@ -1,3 +1,0 @@
-import AuthHeader from './AuthHeader';
-
-export default AuthHeader;
