--- src/components/Auth/@subComponents/AuthContainer/index.ts
@@ -1,3 +1,0 @@
-import AuthContainer from './AuthContainer';
-
-export default AuthContainer;
