--- src/components/Auth/@subComponents/AuthGuide/AuthGuide.type.ts
@@ -1,7 +1,0 @@
-'use client';
-export namespace Props {
-    export interface AuthGuide {
-        title: string;
-        description: string;
-    }
-}
