--- src/components/Auth/@subComponents/AuthGuide/index.ts
@@ -1,3 +1,0 @@
-import AuthGuide from './AuthGuide';
-
-export default AuthGuide;
