--- src/components/Auth/@subComponents/AuthenticationMessage/AuthenticationMessage.type.ts
@@ -1,7 +1,0 @@
-export namespace Props {
-    export interface AuthenticationMessage {
-        suggestedText: string;
-        actionText: string;
-        moveHandler: (e?: React.MouseEvent<HTMLDivElement, MouseEvent>) => void;
-    }
-}
