--- src/components/Auth/@subComponents/AuthenticationMessage/index.ts
@@ -1,3 +1,0 @@
-import AuthenticationMessage from './AuthenticationMessage';
-
-export default AuthenticationMessage;
