--- src/components/Auth/@subComponents/OAuthButton/OAuthButton.type.ts
@@ -1,7 +1,0 @@
-export namespace Props {
-    export interface OAuthButton {
-        src: string;
-        bgColor: string;
-        handler: () => void;
-    }
-}
