--- src/components/Auth/@subComponents/index.ts
@@ -1,4 +1,0 @@
-export { default as OAuthButton } from './OAuthButton';
-export { default as AuthContainer } from './AuthContainer';
-export { default as AuthenticationMessage } from './AuthenticationMessage';
-export { default as AuthGuide } from './AuthGuide';
