+++ src/components/Auth/AuthHeader/AuthHeader.style.tsx
