+++ src/components/Auth/index.ts
@@ -1,1 +1,6 @@
-export { default as ServiceInfoContainer } from './ServiceInfoContainer';
+export { default as AuthContainer } from './AuthContainer/AuthContainer';
+export { default as AuthenticationMessage } from './AuthenticationMessage/AuthenticationMessage';
+export { default as AuthGuide } from './AuthGuide/AuthGuide';
+export { default as AuthHeader } from './AuthHeader/AuthHeader';
+export { default as OAuthButton } from './OAuthButton/OAuthButton';
+export { default as ServiceInfoContainer } from './ServiceInfoContainer/ServiceInfoContainer';
