--- src/components/Landing/@subComponents/ExplaneCard/ExplaneCard.type.ts
@@ -1,8 +1,0 @@
-export namespace Props {
-    export interface ExplaneCard {
-        title: string;
-        description: string;
-        trasitionDelay: number;
-        icon: React.ComponentType<any>;
-    }
-}
