--- src/components/Landing/@subComponents/PricingCard/PricingCard.type.ts
@@ -1,13 +1,0 @@
-export namespace Props {
-    export interface PricingCard {
-        type: string;
-        description: string;
-        benefits: string[];
-        price: string;
-        disabled: boolean;
-        animationDelay: number;
-        moveMainPage: (
-            e?: React.MouseEvent<HTMLButtonElement, MouseEvent>
-        ) => void;
-    }
-}
