--- src/components/Landing/@subComponents/index.ts
@@ -1,2 +1,0 @@
-export { default as ExplaneCard } from './ExplaneCard';
-export { default as PricingCard } from './PricingCard';
