+++ src/components/Landing/ExplaneCard/ExplaneCard.style.tsx
