--- src/components/Landing/LandingContainer/index.ts
@@ -1,3 +1,0 @@
-import LandingContainer from './LandingContainer';
-
-export default LandingContainer;
