+++ src/components/Landing/index.ts
@@ -1,1 +1,3 @@
-export { default as LandingContainer } from './LandingContainer';
+export { default as ExplaneCard } from './ExplaneCard/ExplaneCard';
+export { default as PricingCard } from './PricingCard/PricingCard';
+export { default as LandingContainer } from './LandingContainer/LandingContainer';
