--- src/components/Main/@subComponents/ActionButton/ActionButton.type.ts
@@ -1,5 +1,0 @@
-export namespace Props {
-    export interface ActionButton {
-        title: string;
-    }
-}
