--- src/components/Main/@subComponents/ActionButton/index.ts
@@ -1,3 +1,0 @@
-import ActionButton from './ActionButton';
-
-export default ActionButton;
