--- src/components/Main/@subComponents/NavigationButton/NavigationButton.type.ts
@@ -1,6 +1,0 @@
-export namespace Props {
-    export interface NavigationButton {
-        title: string;
-        urn: string;
-    }
-}
