--- src/components/Main/@subComponents/NavigationButton/index.ts
@@ -1,3 +1,0 @@
-import NavigationButton from './NavigationButton';
-
-export default NavigationButton;
