--- src/components/Main/@subComponents/index.ts
@@ -1,2 +1,0 @@
-export { default as NavigationButton } from './NavigationButton';
-export { default as ActionButton } from './ActionButton';
