+++ src/components/Main/NavigationButton/NavigationButton.tsx
@@ -8,9 +8,13 @@ import { Typography } from 'nimble-ds';
 // emotion
 import { navigationButtonStyle } from './NavigationButton.style';
 
-import type { Props } from './NavigationButton.type';
+// Props type
+export interface Props {
+    title: string;
+    urn: string;
+}
 
-const NavigationButton = ({ title, urn }: Props.NavigationButton) => {
+const NavigationButton = ({ title, urn }: Props) => {
     return (
         <Link href={urn} css={css(navigationButtonStyle)}>
             <Typography value={title} />
