+++ src/components/Main/index.ts
@@ -1,3 +1,5 @@
-export { default as CurrentTimeContainer } from './CurrentTimeContainer';
-export { default as MainContainer } from './MainContainer';
-export { default as Navigation } from './Navigation';
+export { default as ActionButton } from './ActionButton/ActionButton';
+export { default as CurrentTimeContainer } from './CurrentTimeContainer/CurrentTimeContainer';
+export { default as MainContainer } from './MainContainer/MainContainer';
+export { default as Navigation } from './Navigation/Navigation';
+export { default as NavigationButton } from './NavigationButton/NavigationButton';
