--- src/components/Ui/Devider/Devider.type.ts
@@ -1,6 +1,0 @@
-export namespace Props {
-    export interface Devider {
-        width: number;
-        value?: string;
-    }
-}
