--- src/components/Ui/Devider/index.ts
@@ -1,3 +1,0 @@
-import Devider from './Devider';
-
-export default Devider;
