--- src/components/Ui/Footer/index.ts
@@ -1,3 +1,0 @@
-import Footer from './Footer';
-
-export default Footer;
