--- src/components/Ui/Header/index.ts
@@ -1,3 +1,0 @@
-import Header from './Header';
-
-export default Header;
