+++ src/components/Ui/index.ts
@@ -1,4 +1,4 @@
 // common
-export { default as Header } from './Header';
-export { default as Footer } from './Footer';
-export { default as Devider } from './Devider';
+export { default as Header } from './Header/Header';
+export { default as Footer } from './Footer/Footer';
+export { default as Devider } from './Devider/Devider';
